requirements.txt
Django==2.2.8
djangorestframework
djangorestframework-jwt
cryptography
pyjwt
python-dotenv
whitenoise==3.3.1
django-rest-swagger
dj-database-url
psycopg2-binary
requests